+++ CHANGELOG.md
@@ -4,8 +4,11 @@
 
 ### 0.2.0
 
+- Only install tomlkit for Python version less than 3.11 ([#5])
+- Migrate lint tool from isort+black to ruff ([#5])
 - Drop support for Python3.8 ([#4])
 
+[#5]: https://github.com/tortoise/tortoise-cli/pull/5
 [#4]: https://github.com/tortoise/tortoise-cli/pull/4
 
 ## 0.1